change stepTime in cycleStep to double, change aggregation variables to double, adapt precision of testcase for m9

+++ VECTOAux/VectoAuxiliaries/DownstreamModules/IM10.vb
@@ -22,7 +22,7 @@ Namespace DownstreamModules
 		'Interpolated FC between points 2-3-1 Representing smart Pneumatics = Fuel consumption with smart Pneumatics and average electrical  power demand
 		ReadOnly Property FuelConsumptionSmartPneumaticsAndAverageElectricalPowerDemand As Single
 
-		Sub CycleStep(Optional stepTimeInSeconds As Single = 0.0)
+		Sub CycleStep(Optional stepTimeInSeconds As Double = 0.0)
 
 		'Added for diagnostic inspection purposes only, does not materially affect the class function.
 		ReadOnly Property P1X As Single
